+++ src/vips.lua
@@ -7,8 +7,6 @@ local vips = ffi.load(ffi.os == "Windows" and "libvips-42.dll" or "vips")
 ffi.cdef[[
     int vips_init (const char* argv0);
 
-    const char* vips_error_buffer (void);
-    void vips_error_clear (void);
 ]]
 
 local result = vips.vips_init("")
@@ -20,6 +18,8 @@ if result ~= 0 then
 end
 
 return { 
+    verror = require "vips/verror",
+    version = require "vips/version",
     log = require "vips/log",
     gvalue = require "vips/gvalue",
     vobject = require "vips/vobject",