+++ src/vips/verror.lua
@@ -1,0 +1,27 @@
+-- handle the libvips error buffer
+
+local ffi = require "ffi" 
+
+local vips = ffi.load(ffi.os == "Windows" and "libvips-42.dll" or "vips")
+
+ffi.cdef[[
+    const char* vips_error_buffer (void);
+    void vips_error_clear (void);
+
+]]
+
+local verror = {}
+
+verror = {
+    -- get and clear the error buffer 
+    get = function()
+        local errstr = ffi.string(vips.vips_error_buffer())
+        vips.vips_error_clear()
+
+        return errstr
+    end,
+
+}
+
+return verror
+